+++ src/main/java/se/lexicon/model/ShippingRequest.java
@@ -1,5 +1,7 @@
 package se.lexicon.model;
 
 
-public record ShippingRequest(Destination destination, Speed speed, double weightKg) {
+import java.math.BigDecimal;
+
+public record ShippingRequest(Destination destination, Speed speed, BigDecimal weightKg) {
 }
