+++ src/main/java/se/lexicon/service/ShippingCostCalculator.java
@@ -2,10 +2,12 @@ package se.lexicon.service;
 
 import se.lexicon.model.ShippingRequest;
 
+import java.math.BigDecimal;
+
 public interface ShippingCostCalculator {
     boolean supports(ShippingRequest req);
 
-    double calculate(ShippingRequest req);
+    BigDecimal calculate(ShippingRequest req);
 }
