+++ src/main/java/se/lexicon/service/ShippingService.java
@@ -3,6 +3,8 @@ package se.lexicon.service;
 import org.springframework.stereotype.Service;
 import se.lexicon.model.ShippingRequest;
 
+import java.math.BigDecimal;
+
 @Service
 public class ShippingService {
     private final ShippingCalculatorFactory factory;
@@ -11,7 +13,7 @@ public class ShippingService {
         this.factory = factory;
     }
 
-    public double quote(ShippingRequest req) {
+    public BigDecimal quote(ShippingRequest req) {
         ShippingCostCalculator calc = factory.getCalculator(req);
         return calc.calculate(req);
     }
